REL.00.01.00.02. Compatible with: J7 PSDK-RTOS 07_01_00_09, tidl-semseg-model_1.3.0.11.tar.gz, ros-bag_2020_0914.tar.gz

+++ nodes/ti_sde/src/sde.h
@@ -194,7 +194,7 @@ typedef struct
     uint8_t                                inputFormat;
 
     /** Application interactive status, 0=non-interactive, 1=interactive. */
-    uint32_t                               is_interactive;
+    uint8_t                                is_interactive;
 
     SDELDCAPPLIB_createParams              sdeLdcCreateParams;
     SDELDCAPPLIB_Handle                    sdeLdcHdl;
